requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-identity==1.15.0
pandas==2.2.0
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pydantic==2.6.1
python-dotenv==1.0.1
jinja2==3.1.3 